refactor : RacingCarGameConfig 상수 패키지로 이동
