+++ src/main/java/racingcar/constant/RacingCarGameConfig.java
@@ -1,10 +1,9 @@
-package racingcar.model;
+package racingcar.constant;
 
 public class RacingCarGameConfig {
 
   public static final Integer CARNAME_MAXIMUM_LENGTH = 5;
   public static final Integer RANDOM_MINIMUM_VALUE = 0;
   public static final Integer RANDOM_MAXIMUM_VALUE = 9;
-
   public static final Integer GAME_WIN_CONDITION = 4;
 }
